refactor: move page component into page folder
